Remove debug assertion
Since Geany can be configured to have no documents open, instead of
opening a new file when the last file is closed. Basically, if there's
no document open, since there's no way to check, it assumes putting
the overview bar on the left is not supported, even if it is/will be.

TODO: make it a build-time check rather than runtime, using Geany's
API version number greater than current.